Refactor vectordb into subpackages
